+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+### 1.2.0 (2020-04-03)
+  1. **New features**
+        1. Now supports multiple currencies per store
+        2. Product custom options are now extracted
+  2. **Improvements**
+        1. Product category output (now returned as a category tree)
+        2. Reduced database queries
+        3. Reduced memory usage
+  
 ### 1.1.8 (2020-04-01)
   1. Fixed an issue related to variant product options.